+++ R/DistMatDR-Package.R
@@ -2,7 +2,7 @@
 #'
 #' @name DistMatDR - package
 #' @docType package
-#' @title Distance Matrix of traveltime on foot for the 'German Reich'
+#' @title Distance Matrix of traveltime by foot for the 'German Reich'
 #' @description The \code{DistMatDR} package is used to get the traveltime on foot between desired GIDs in the 'German Reich'.
 #' @author Andreas Schönberg
 #' @keywords package